[composer] build

+++ include/global.php
@@ -23475,6 +23475,7 @@ if (!isset($excluded_functions["process_async"]) && (!function_exists("process_a
             private $proc;
             private $pipes;
             private $status;
+            private $destructAction;
             public  $stdout;
             public  $stderr;
 
@@ -23484,19 +23485,25 @@ if (!isset($excluded_functions["process_async"]) && (!function_exists("process_a
                 $this->pipes = $pipes;
                 $this->stdout = &$stdout;
                 $this->stderr = &$stderr;
+                $this->destructAction = 'close';
             }
 
             public function __destruct()
             {
-                if ($this->proc !== null) {
-                    fclose($this->pipes[1]);
-                    fclose($this->pipes[2]);
-                    proc_close($this->proc);
+                if ($this->destructAction === 'close') {
+                    $this->__invoke();
+                }
+                if ($this->destructAction === 'terminate') {
+                    $this->terminate();
                 }
             }
 
             public function __invoke()
             {
+                if ($this->proc === null) {
+                    return $this->status['exitcode'];
+                }
+
                 try {
                     /** @noinspection PhpStatementHasEmptyBodyInspection */
                     while ($this->update()) {
@@ -23512,6 +23519,12 @@ if (!isset($excluded_functions["process_async"]) && (!function_exists("process_a
                 }
 
                 return $this->status['running'] ? $rc : $this->status['exitcode'];
+            }
+
+            public function setDestructAction($action)
+            {
+                $this->destructAction = $action;
+                return $this;
             }
 
             public function update()
@@ -23556,6 +23569,10 @@ if (!isset($excluded_functions["process_async"]) && (!function_exists("process_a
 
             public function terminate()
             {
+                if ($this->proc === null) {
+                    return !$this->status['running'];
+                }
+
                 fclose($this->pipes[1]);
                 fclose($this->pipes[2]);
                 proc_terminate($this->proc);
